+++ main.py
@@ -5,4 +5,3 @@ def main():
 if __name__ == "__main__":
     main()
 
-
