+++ shared-config/shared_values.txt
@@ -1,0 +1,4 @@
+Voltage: 1
+Temperature: 
+Fluence Min: 
+Fluence Max:
