--- testdata.csv
@@ -1,81 +1,0 @@
-Delta Ib,Current
-403851000000.0,0.000298364014
-432938000000.0,0.000298290915
-464079500000.0,0.000298214911
-497851000000.0,0.000298130176
-533835000000.0,0.000298049501
-572118500000.0,0.000297951544
-613795500000.0,0.000297859896
-657966500000.0,0.000297758201
-705240000000.0,0.000297644754
-756231500000.0,0.00029753017
-811065500000.0,0.000297403823
-869874500000.0,0.000297277316
-932801500000.0,0.000297130251
-1000000000000.0,0.00029698188
-1071135000000.0,0.000296822368
-1149880000000.0,0.000296653168
-1231425000000.0,0.000296472281
-1320970000000.0,0.00029627914
-1418735000000.0,0.000296077153
-1519955000000.0,0.000295859749
-1629875000000.0,0.000295626153
-1748765000000.0,0.000295380934
-1871910000000.0,0.000295113493
-2009615000000.0,0.000294835167
-2152215000000.0,0.000294525874
-2310065000000.0,0.000294205897
-2478540000000.0,0.000293860668
-2658045000000.0,0.000293487929
-2849020000000.0,0.000293085717
-3051930000000.0,0.000292652805
-3272275000000.0,0.000292187015
-3510595000000.0,0.000291687393
-3762470000000.0,0.000291127824
-4038510000000.0,0.000290522577
-4329380000000.0,0.000289827256
-4640795000000.0,0.00028903794
-4978510000000.0,0.000287926915
-5338350000000.0,0.000249624808
-5721185000000.0,0.000238156613
-6137955000000.0,0.000226614874
-6579665000000.0,0.000216853025
-7052400000000.0,0.000205918855
-7562315000000.0,0.000196842283
-8110655000000.0,0.000187650455
-8698745000000.0,0.000179421228
-9328015000000.0,0.000171158533
-10000000000000.0,0.000162934113
-10711350000000.0,0.000155929143
-11498800000000.0,0.000150362536
-12314250000000.0,0.000146569028
-13209700000000.0,0.000143977112
-14187350000000.0,0.00014150546
-15199550000000.0,0.000138944008
-16298750000000.0,0.000136116092
-17487650000000.0,0.000133315743
-18719100000000.0,0.000130529702
-20096150000000.0,0.000127840151
-21522150000000.0,0.000125194322
-23100650000000.0,0.000122602107
-24785400000000.0,0.000120148937
-26580450000000.0,0.000117696129
-28490200000000.0,0.000115407174
-30519300000000.0,0.000113134386
-32722750000000.0,0.000110965943
-35105950000000.0,0.000108825554
-37624700000000.0,0.000106805991
-40385100000000.0,0.00010482666
-43293800000000.0,0.000102984782
-46407950000000.0,0.00010119569
-49785100000000.0,9.94634552e-05
-53383500000000.0,9.77923086e-05
-57211850000000.0,9.61865819e-05
-61379550000000.0,9.4650527e-05
-65796650000000.0,9.30830007e-05
-70524000000000.0,9.16974646e-05
-75623150000000.0,9.03947896e-05
-81106550000000.0,8.90679209e-05
-86987450000000.0,8.78289428e-05
-93280150000000.0,8.65667659e-05
-100000000000000.0,8.55155491e-05
